Fix review comments:  merge test code from make_optional.pass.cpp, make_optional_explicit.pass.cpp and make_optional_explicit_initializer_list.pass.cpp into one make_optional.pass.cpp
