--- test/xpu_api/utilities/optional/optional.specalg/make_optional_explicit_initializer_list.pass.cpp
@@ -1,63 +1,0 @@
-// -*- C++ -*-
-//===----------------------------------------------------------------------===//
-//
-// Copyright (C) Intel Corporation
-//
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-// This file incorporates work covered by the following copyright and permission
-// notice:
-//
-// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-//
-//===----------------------------------------------------------------------===//
-
-#include "support/test_config.h"
-
-#include <oneapi/dpl/optional>
-#include <oneapi/dpl/type_traits>
-
-#include "support/test_macros.h"
-#include "support/utils.h"
-
-#if TEST_DPCPP_BACKEND_PRESENT
-struct TestT
-{
-    int x;
-    int size;
-    constexpr TestT(std::initializer_list<int> il) : x(*il.begin()), size(static_cast<int>(il.size())) {}
-    constexpr TestT(std::initializer_list<int> il, const int*) : x(*il.begin()), size(static_cast<int>(il.size())) {}
-};
-
-void
-kernel_test()
-{
-    sycl::queue q = TestUtils::get_test_queue();
-    q.submit([&](sycl::handler& cgh) {
-        cgh.single_task<class KernelTest>([=]() {
-            using dpl::make_optional;
-            {
-                constexpr auto opt = make_optional<TestT>({42, 2, 3});
-                static_assert(opt->x == 42);
-                static_assert(opt->size == 3);
-            }
-            {
-                constexpr auto opt = make_optional<TestT>({42, 2, 3}, nullptr);
-                static_assert(opt->x == 42);
-                static_assert(opt->size == 3);
-            }
-        });
-    });
-}
-#endif // TEST_DPCPP_BACKEND_PRESENT
-
-int
-main()
-{
-#if TEST_DPCPP_BACKEND_PRESENT
-    kernel_test();
-#endif // TEST_DPCPP_BACKEND_PRESENT
-
-    return TestUtils::done(TEST_DPCPP_BACKEND_PRESENT);
-}
